Add a function that evaluates the winner
I added a function called playRound() that takes the input from the player and the computerChoice()
and then evaluates the result. To be honest, the procedure of evaluation is a bit more complicated
than it had to be. But it is my first JavaScript code after all.

Additionally, I made a new text file that is called Pseudocode. It's the rough concept of the
rock, paper, scissors game that I am making.